+++ docs/lib/examples/Tooltip.js
@@ -1,6 +1,8 @@
 /* eslint react/no-multi-comp: 0, react/prop-types: 0 */
 import React from 'react';
 import { Tooltip } from 'reactstrap';
+import { mapToCssModules } from '../../../src/utils'
+import classNames from 'classnames'
 
 export default class Example extends React.Component {
   constructor(props) {
@@ -18,7 +20,10 @@ export default class Example extends React.Component {
     });
   }
 
+
+
   render() {
+    const classes = 'tooltip-inner'
     return (
       <div>
         <p>Somewhere in here is a <span style={{textDecoration: "underline", color:"blue"}} href="#" id="TooltipExample">tooltip</span>.</p>
