+++ src/__tests__/TooltipPopoverWrapper.spec.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import { Tooltip, PopperContent } from '../';
+import { PopperContent } from '../';
+import TooltipPopoverWrapper from '../TooltipPopoverWrapper';
 
 describe('Tooltip', () => {
   let isOpen;
@@ -16,7 +17,7 @@ describe('Tooltip', () => {
     toggle = () => { isOpen = !isOpen; };
     element = document.createElement('div');
     container = document.createElement('div');
-    element.innerHTML = '<p id="target">This is the tooltip <span id="innerTarget">target</span>.</p>';
+    element.innerHTML = '<p id="target">This is the Tooltip <span id="innerTarget">target</span>.</p>';
     element.setAttribute('id', 'testContainer');
     container.setAttribute('id', 'container');
     element.appendChild(container);
@@ -40,93 +41,93 @@ describe('Tooltip', () => {
 
   it('should render with "hideArrow" false by default', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle}>
         Tooltip Content
-      </Tooltip>);
+      </TooltipPopoverWrapper>);
 
     expect(wrapper.prop('hideArrow')).toBe(false);
   });
 
   it('should render with "hideArrow" true when "hideArrow" prop is truthy', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle} hideArrow>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} hideArrow>
         Tooltip Content
-      </Tooltip>);
+      </TooltipPopoverWrapper>);
 
     expect(wrapper.prop('hideArrow')).toBe(true);
   });
 
   it('should not render children if isOpen is false', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
 
-    const tooltips = document.getElementsByClassName('tooltip');
+    const Tooltips = document.getElementsByClassName('Tooltip');
 
-    expect(wrapper.find('.tooltip.show').hostNodes().length).toBe(0);
+    expect(wrapper.find('.Tooltip.show').hostNodes().length).toBe(0);
     expect(target.className).toBe('');
-    expect(tooltips.length).toBe(0);
+    expect(Tooltips.length).toBe(0);
     wrapper.detach();
   });
 
   it('should render if isOpen is true', () => {
     isOpen = true;
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} popperClasses="tooltip show">
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
 
-    const tooltips = document.getElementsByClassName('tooltip');
+    const Tooltips = document.getElementsByClassName('tooltip');
 
-    expect(wrapper.find('.tooltip.show').hostNodes().length).toBe(0);
-    expect(tooltips.length).toBe(1);
-    expect(tooltips[0].textContent).toBe('Tooltip Content');
+    expect(wrapper.find('.Tooltip.show').hostNodes().length).toBe(0);
+    expect(Tooltips.length).toBe(1);
+    expect(Tooltips[0].textContent).toBe('Tooltip Content');
     wrapper.detach();
   });
 
   it('should render with target object', () => {
     isOpen = true;
     const wrapper = mount(
-      <Tooltip target={document.getElementById('target')} isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target={document.getElementById('target')} isOpen={isOpen} toggle={toggle} popperClasses="tooltip show">
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
 
-    const tooltips = document.getElementsByClassName('tooltip');
+    const Tooltips = document.getElementsByClassName('Tooltip');
 
-    expect(wrapper.find('.tooltip.show').hostNodes().length).toBe(0);
-    expect(tooltips.length).toBe(1);
-    expect(tooltips[0].textContent).toBe('Tooltip Content');
+    expect(wrapper.find('.Tooltip.show').hostNodes().length).toBe(0);
+    expect(Tooltips.length).toBe(1);
+    expect(Tooltips[0].textContent).toBe('Tooltip Content');
     wrapper.detach();
   });
 
   it('should toggle isOpen', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} popperClasses="tooltip show">
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
 
-    expect(document.getElementsByClassName('tooltip').length).toBe(0);
+    expect(document.getElementsByClassName('Tooltip').length).toBe(0);
     wrapper.setProps({ isOpen: true });
-    expect(document.getElementsByClassName('tooltip').length).toBe(1);
+    expect(document.getElementsByClassName('Tooltip').length).toBe(1);
     wrapper.setProps({ isOpen: false });
-    expect(document.getElementsByClassName('tooltip').length).toBe(0);
+    expect(document.getElementsByClassName('Tooltip').length).toBe(0);
     wrapper.detach();
   });
 
   it('should handle target clicks', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
     const instance = wrapper.instance();
@@ -142,9 +143,9 @@ describe('Tooltip', () => {
 
   it('should handle inner target clicks', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
     const instance = wrapper.instance();
@@ -157,9 +158,9 @@ describe('Tooltip', () => {
 
   it('should not do anything when document click outside of target', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
     const instance = wrapper.instance();
@@ -173,9 +174,9 @@ describe('Tooltip', () => {
 
   it('should clear hide timeout if it exists on target click', () => {
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen} toggle={toggle} delay={200}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} delay={200}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
     const instance = wrapper.instance();
@@ -192,33 +193,9 @@ describe('Tooltip', () => {
     wrapper.detach();
   });
 
-  it('should allow custom classes to be added to the tooltip-inner', () => {
-    const wrapper = mount(
-      <Tooltip isOpen target="target" innerClassName="tooltip-special">
-        Tooltip Content
-      </Tooltip>
-    );
-
-    expect(document.getElementsByClassName('tooltip-inner')[0].className.indexOf('tooltip-special') > -1).toBe(true);
-
-    wrapper.unmount();
-  });
-
-  it('should allow custom classes to be added to the tooltip', () => {
-    const wrapper = mount(
-      <Tooltip isOpen target="target" className="tooltip-special">
-        Tooltip Content
-      </Tooltip>
-    );
-
-    expect(document.getElementsByClassName('tooltip')[0].className.indexOf('tooltip-special') > -1).toBe(true);
-
-    wrapper.unmount();
-  });
-
   it('should pass down custom modifiers', () => {
     const wrapper = mount(
-      <Tooltip
+      <TooltipPopoverWrapper
         isOpen
         target="target"
         modifiers={{
@@ -227,7 +204,7 @@ describe('Tooltip', () => {
         }}
       >
         Tooltip Content
-      </Tooltip>
+      </TooltipPopoverWrapper>
     );
 
     expect(wrapper.find(PopperContent).props().modifiers).toEqual({
@@ -241,13 +218,13 @@ describe('Tooltip', () => {
   it('should pass down cssModule', () => {
     const cssModule = {};
     const wrapper = mount(
-      <Tooltip
+      <TooltipPopoverWrapper
         isOpen
         target="target"
         cssModule={cssModule}
       >
         Tooltip Content
-      </Tooltip>
+      </TooltipPopoverWrapper>
     );
     expect(wrapper.find(PopperContent).props().cssModule).toBe(cssModule);
     wrapper.unmount();
@@ -255,9 +232,9 @@ describe('Tooltip', () => {
 
   it('should pass down offset', () => {
     const wrapper = mount(
-      <Tooltip isOpen target="target" offset="100">
+      <TooltipPopoverWrapper isOpen target="target" offset="100">
         Tooltip content
-      </Tooltip>
+      </TooltipPopoverWrapper>
     );
 
     expect(wrapper.find(PopperContent).props().offset).toEqual('100');
@@ -269,9 +246,9 @@ describe('Tooltip', () => {
     const event = createSpyObj('event', ['preventDefault']);
 
     const wrapper = mount(
-      <Tooltip target="target" disabled isOpen={isOpen} toggle={props.toggle}>
+      <TooltipPopoverWrapper target="target" disabled isOpen={isOpen} toggle={props.toggle}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
     const instance = wrapper.instance();
@@ -288,9 +265,9 @@ describe('Tooltip', () => {
     const event = createSpyObj('event', ['preventDefault']);
 
     const wrapper = mount(
-      <Tooltip target="target" isOpen={isOpen}>
+      <TooltipPopoverWrapper target="target" isOpen={isOpen}>
         Tooltip Content
-      </Tooltip>,
+      </TooltipPopoverWrapper>,
       { attachTo: container }
     );
     const instance = wrapper.instance();
@@ -304,9 +281,9 @@ describe('Tooltip', () => {
     it('should accept a number', () => {
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={toggle} delay={200}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -320,9 +297,9 @@ describe('Tooltip', () => {
     it('should accept an object', () => {
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={toggle} delay={{ show: 200, hide: 200 }}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} delay={{ show: 200, hide: 200 }}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -336,9 +313,9 @@ describe('Tooltip', () => {
     it('should use default value if value is missing from object', () => {
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={toggle} delay={{ show: 0 }}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={toggle} delay={{ show: 0 }}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -355,9 +332,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -374,9 +351,9 @@ describe('Tooltip', () => {
     it('should not call toggle when isOpen is false', () => {
       const spy = jest.fn(toggle);
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -395,9 +372,9 @@ describe('Tooltip', () => {
     it('should call toggle when isOpen is false', () => {
       const spy = jest.fn(toggle);
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -415,9 +392,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -436,9 +413,9 @@ describe('Tooltip', () => {
     it('should clear timeout if it exists on target click', () => {
       const spy = jest.fn(toggle);
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy} delay={200}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -460,9 +437,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy} delay={0}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy} delay={0}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -482,9 +459,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy} delay={200}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -506,9 +483,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = false;
       const wrapper = mount(
-        <Tooltip target="target" isOpen={isOpen} toggle={spy} delay={0}>
+        <TooltipPopoverWrapper target="target" isOpen={isOpen} toggle={spy} delay={0}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -524,13 +501,13 @@ describe('Tooltip', () => {
   });
 
   describe('autohide', () => {
-    it('should keep tooltip around when false and onmouseleave from tooltip content', () => {
+    it('should keep Tooltip around when false and onmouseleave from Tooltip content', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" autohide={false} isOpen={isOpen} toggle={spy} delay={200}>
+        <TooltipPopoverWrapper target="target" autohide={false} isOpen={isOpen} toggle={spy} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -551,9 +528,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" autohide={false} isOpen={isOpen} toggle={spy} delay={200}>
+        <TooltipPopoverWrapper target="target" autohide={false} isOpen={isOpen} toggle={spy} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -570,9 +547,9 @@ describe('Tooltip', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" autohide={false} isOpen={isOpen} toggle={spy} delay={200}>
+        <TooltipPopoverWrapper target="target" autohide={false} isOpen={isOpen} toggle={spy} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
@@ -588,13 +565,13 @@ describe('Tooltip', () => {
       wrapper.detach();
     });
 
-    it('should not keep tooltip around when autohide is true and tooltip content is hovered over', () => {
+    it('should not keep Tooltip around when autohide is true and Tooltip content is hovered over', () => {
       const spy = jest.fn(toggle);
       isOpen = true;
       const wrapper = mount(
-        <Tooltip target="target" autohide isOpen={isOpen} toggle={spy} delay={200}>
+        <TooltipPopoverWrapper target="target" autohide isOpen={isOpen} toggle={spy} delay={200}>
           Tooltip Content
-        </Tooltip>,
+        </TooltipPopoverWrapper>,
         { attachTo: container }
       );
       const instance = wrapper.instance();
